Require bindgen updates for wepoll source changes
When https://gitlab.com/yorickpeterse/wepoll-sys/-/merge_requests/2 is
merged this will ensure that any change to the wepoll/ directory also
requires a change to the generated bindings. Technically it's possible
for a source change that does not result in a change for the generated
bindings. We'll deal with such cases when they present themselves.