Remove bindgen dependency
Updating the bindgen bindings didn't work reliably due to it writing
output to the wrong directory. The linter for bindgen in turn would
produce false positives.

Since the wepoll API doesn't change often, there's a simple solution:
just maintain the bindings manually, and get rid of bindgen. This
removes the need for the linter, and simplifies building and maintaining
this crate.